refactor: 랜덤으로 Line 을 생성하는 로직을 클래스로 분리
