+++ src/main/java/ladder/domain/creator/RandomLadderCreator.java
@@ -1,0 +1,19 @@
+package ladder.domain.creator;
+
+import ladder.domain.Ladder;
+
+import java.util.stream.IntStream;
+
+public class RandomLadderCreator {
+    private final RandomLineCreator randomLineCreator;
+
+    public RandomLadderCreator(RandomLineCreator randomLineCreator) {
+        this.randomLineCreator = randomLineCreator;
+    }
+
+    public Ladder create(int width, int height) {
+        return new Ladder(IntStream.range(0, height)
+                .mapToObj(i -> randomLineCreator.create(width))
+                .toList());
+    }
+}
